refactor(browser): dedupe browser and cli command wiring
(cherry picked from commit 067855e623)
