+++ src/cli/browser-cli-actions-input/register.element.ts
@@ -2,12 +2,42 @@ import type { Command } from "commander";
 import { danger } from "../../globals.js";
 import { defaultRuntime } from "../../runtime.js";
 import type { BrowserParentOpts } from "../browser-cli-shared.js";
-import { callBrowserAct, requireRef, resolveBrowserActionContext } from "./shared.js";
+import {
+  callBrowserAct,
+  logBrowserActionResult,
+  requireRef,
+  resolveBrowserActionContext,
+} from "./shared.js";
 
 export function registerBrowserElementCommands(
   browser: Command,
   parentOpts: (cmd: Command) => BrowserParentOpts,
 ) {
+  const runElementAction = async (params: {
+    cmd: Command;
+    body: Record<string, unknown>;
+    successMessage: string | ((result: unknown) => string);
+    timeoutMs?: number;
+  }): Promise<void> => {
+    const { parent, profile } = resolveBrowserActionContext(params.cmd, parentOpts);
+    try {
+      const result = await callBrowserAct({
+        parent,
+        profile,
+        body: params.body,
+        timeoutMs: params.timeoutMs,
+      });
+      const successMessage =
+        typeof params.successMessage === "function"
+          ? params.successMessage(result)
+          : params.successMessage;
+      logBrowserActionResult(parent, result, successMessage);
+    } catch (err) {
+      defaultRuntime.error(danger(String(err)));
+      defaultRuntime.exit(1);
+    }
+  };
+
   browser
     .command("click")
     .description("Click an element by ref from snapshot")
@@ -17,7 +47,6 @@ export function registerBrowserElementCommands(
     .option("--button <left|right|middle>", "Mouse button to use")
     .option("--modifiers <list>", "Comma-separated modifiers (Shift,Alt,Meta)")
     .action(async (ref: string | undefined, opts, cmd) => {
-      const { parent, profile } = resolveBrowserActionContext(cmd, parentOpts);
       const refValue = requireRef(ref);
       if (!refValue) {
         return;
@@ -28,29 +57,22 @@ export function registerBrowserElementCommands(
             .map((v: string) => v.trim())
             .filter(Boolean)
         : undefined;
-      try {
-        const result = await callBrowserAct<{ url?: string }>({
-          parent,
-          profile,
-          body: {
-            kind: "click",
-            ref: refValue,
-            targetId: opts.targetId?.trim() || undefined,
-            doubleClick: Boolean(opts.double),
-            button: opts.button?.trim() || undefined,
-            modifiers,
-          },
-        });
-        if (parent?.json) {
-          defaultRuntime.log(JSON.stringify(result, null, 2));
-          return;
-        }
-        const suffix = result.url ? ` on ${result.url}` : "";
-        defaultRuntime.log(`clicked ref ${refValue}${suffix}`);
-      } catch (err) {
-        defaultRuntime.error(danger(String(err)));
-        defaultRuntime.exit(1);
-      }
+      await runElementAction({
+        cmd,
+        body: {
+          kind: "click",
+          ref: refValue,
+          targetId: opts.targetId?.trim() || undefined,
+          doubleClick: Boolean(opts.double),
+          button: opts.button?.trim() || undefined,
+          modifiers,
+        },
+        successMessage: (result) => {
+          const url = (result as { url?: unknown }).url;
+          const suffix = typeof url === "string" && url ? ` on ${url}` : "";
+          return `clicked ref ${refValue}${suffix}`;
+        },
+      });
     });
 
   browser
@@ -62,33 +84,22 @@ export function registerBrowserElementCommands(
     .option("--slowly", "Type slowly (human-like)", false)
     .option("--target-id <id>", "CDP target id (or unique prefix)")
     .action(async (ref: string | undefined, text: string, opts, cmd) => {
-      const { parent, profile } = resolveBrowserActionContext(cmd, parentOpts);
       const refValue = requireRef(ref);
       if (!refValue) {
         return;
       }
-      try {
-        const result = await callBrowserAct({
-          parent,
-          profile,
-          body: {
-            kind: "type",
-            ref: refValue,
-            text,
-            submit: Boolean(opts.submit),
-            slowly: Boolean(opts.slowly),
-            targetId: opts.targetId?.trim() || undefined,
-          },
-        });
-        if (parent?.json) {
-          defaultRuntime.log(JSON.stringify(result, null, 2));
-          return;
-        }
-        defaultRuntime.log(`typed into ref ${refValue}`);
-      } catch (err) {
-        defaultRuntime.error(danger(String(err)));
-        defaultRuntime.exit(1);
-      }
+      await runElementAction({
+        cmd,
+        body: {
+          kind: "type",
+          ref: refValue,
+          text,
+          submit: Boolean(opts.submit),
+          slowly: Boolean(opts.slowly),
+          targetId: opts.targetId?.trim() || undefined,
+        },
+        successMessage: `typed into ref ${refValue}`,
+      });
     });
 
   browser
@@ -97,22 +108,11 @@ export function registerBrowserElementCommands(
     .argument("<key>", "Key to press (e.g. Enter)")
     .option("--target-id <id>", "CDP target id (or unique prefix)")
     .action(async (key: string, opts, cmd) => {
-      const { parent, profile } = resolveBrowserActionContext(cmd, parentOpts);
-      try {
-        const result = await callBrowserAct({
-          parent,
-          profile,
-          body: { kind: "press", key, targetId: opts.targetId?.trim() || undefined },
-        });
-        if (parent?.json) {
-          defaultRuntime.log(JSON.stringify(result, null, 2));
-          return;
-        }
-        defaultRuntime.log(`pressed ${key}`);
-      } catch (err) {
-        defaultRuntime.error(danger(String(err)));
-        defaultRuntime.exit(1);
-      }
+      await runElementAction({
+        cmd,
+        body: { kind: "press", key, targetId: opts.targetId?.trim() || undefined },
+        successMessage: `pressed ${key}`,
+      });
     });
 
   browser
@@ -121,22 +121,11 @@ export function registerBrowserElementCommands(
     .argument("<ref>", "Ref id from snapshot")
     .option("--target-id <id>", "CDP target id (or unique prefix)")
     .action(async (ref: string, opts, cmd) => {
-      const { parent, profile } = resolveBrowserActionContext(cmd, parentOpts);
-      try {
-        const result = await callBrowserAct({
-          parent,
-          profile,
-          body: { kind: "hover", ref, targetId: opts.targetId?.trim() || undefined },
-        });
-        if (parent?.json) {
-          defaultRuntime.log(JSON.stringify(result, null, 2));
-          return;
-        }
-        defaultRuntime.log(`hovered ref ${ref}`);
-      } catch (err) {
-        defaultRuntime.error(danger(String(err)));
-        defaultRuntime.exit(1);
-      }
+      await runElementAction({
+        cmd,
+        body: { kind: "hover", ref, targetId: opts.targetId?.trim() || undefined },
+        successMessage: `hovered ref ${ref}`,
+      });
     });
 
   browser
@@ -148,32 +137,22 @@ export function registerBrowserElementCommands(
       Number(v),
     )
     .action(async (ref: string | undefined, opts, cmd) => {
-      const { parent, profile } = resolveBrowserActionContext(cmd, parentOpts);
       const refValue = requireRef(ref);
       if (!refValue) {
         return;
       }
-      try {
-        const result = await callBrowserAct({
-          parent,
-          profile,
-          body: {
-            kind: "scrollIntoView",
-            ref: refValue,
-            targetId: opts.targetId?.trim() || undefined,
-            timeoutMs: Number.isFinite(opts.timeoutMs) ? opts.timeoutMs : undefined,
-          },
-          timeoutMs: Number.isFinite(opts.timeoutMs) ? opts.timeoutMs : undefined,
-        });
-        if (parent?.json) {
-          defaultRuntime.log(JSON.stringify(result, null, 2));
-          return;
-        }
-        defaultRuntime.log(`scrolled into view: ${refValue}`);
-      } catch (err) {
-        defaultRuntime.error(danger(String(err)));
-        defaultRuntime.exit(1);
-      }
+      const timeoutMs = Number.isFinite(opts.timeoutMs) ? opts.timeoutMs : undefined;
+      await runElementAction({
+        cmd,
+        body: {
+          kind: "scrollIntoView",
+          ref: refValue,
+          targetId: opts.targetId?.trim() || undefined,
+          timeoutMs,
+        },
+        timeoutMs,
+        successMessage: `scrolled into view: ${refValue}`,
+      });
     });
 
   browser
@@ -183,27 +162,16 @@ export function registerBrowserElementCommands(
     .argument("<endRef>", "End ref id")
     .option("--target-id <id>", "CDP target id (or unique prefix)")
     .action(async (startRef: string, endRef: string, opts, cmd) => {
-      const { parent, profile } = resolveBrowserActionContext(cmd, parentOpts);
-      try {
-        const result = await callBrowserAct({
-          parent,
-          profile,
-          body: {
-            kind: "drag",
-            startRef,
-            endRef,
-            targetId: opts.targetId?.trim() || undefined,
-          },
-        });
-        if (parent?.json) {
-          defaultRuntime.log(JSON.stringify(result, null, 2));
-          return;
-        }
-        defaultRuntime.log(`dragged ${startRef} → ${endRef}`);
-      } catch (err) {
-        defaultRuntime.error(danger(String(err)));
-        defaultRuntime.exit(1);
-      }
+      await runElementAction({
+        cmd,
+        body: {
+          kind: "drag",
+          startRef,
+          endRef,
+          targetId: opts.targetId?.trim() || undefined,
+        },
+        successMessage: `dragged ${startRef} → ${endRef}`,
+      });
     });
 
   browser
@@ -213,26 +181,15 @@ export function registerBrowserElementCommands(
     .argument("<values...>", "Option values to select")
     .option("--target-id <id>", "CDP target id (or unique prefix)")
     .action(async (ref: string, values: string[], opts, cmd) => {
-      const { parent, profile } = resolveBrowserActionContext(cmd, parentOpts);
-      try {
-        const result = await callBrowserAct({
-          parent,
-          profile,
-          body: {
-            kind: "select",
-            ref,
-            values,
-            targetId: opts.targetId?.trim() || undefined,
-          },
-        });
-        if (parent?.json) {
-          defaultRuntime.log(JSON.stringify(result, null, 2));
-          return;
-        }
-        defaultRuntime.log(`selected ${values.join(", ")}`);
-      } catch (err) {
-        defaultRuntime.error(danger(String(err)));
-        defaultRuntime.exit(1);
-      }
+      await runElementAction({
+        cmd,
+        body: {
+          kind: "select",
+          ref,
+          values,
+          targetId: opts.targetId?.trim() || undefined,
+        },
+        successMessage: `selected ${values.join(", ")}`,
+      });
     });
 }
