+++ src/cli/browser-cli-actions-input/shared.ts
@@ -40,6 +40,18 @@ export async function callBrowserAct<T = unknown>(params: {
   );
 }
 
+export function logBrowserActionResult(
+  parent: BrowserParentOpts,
+  result: unknown,
+  successMessage: string,
+) {
+  if (parent?.json) {
+    defaultRuntime.log(JSON.stringify(result, null, 2));
+    return;
+  }
+  defaultRuntime.log(successMessage);
+}
+
 export function requireRef(ref: string | undefined) {
   const refValue = typeof ref === "string" ? ref.trim() : "";
   if (!refValue) {
